+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ### Added
 
+* `<EdgeLabel />`: use `title` property on its text sub element
 * `<Application* />` elements now have defined and exposed interfaces
 * `<FieldSet />` element now have a defined and exposed interface
 * `<PropertyValue* />` elements now have defined and exposed interfaces